fix(mir): stop reporting an unresolved place for a poisoned binding's assignment
When a let/var initializer fails to lower, the binding never gets an MIR
place and its failure is recorded in poisoned_let_bindings so a later
READ of the binding stays silent instead of stacking a second diagnostic.
The assignment-target path had no equivalent guard, so reassigning a
poisoned binding produced a second, misleading UnresolvedPlace error on
top of the real root cause.

Add the same poison check to the assignment arm, immediately before the
UnresolvedPlace push, so it only suppresses the cascade for a binding
already recorded as poisoned by a genuine earlier failure. A binding
that is genuinely unresolved and was never poisoned still reports.

Also splits the existing binding_ty_is_plain_vec_tuple test module and
the new poisoned_assign_target_cascade test module out of expr.rs into
sibling files under src/lower/expr/ (pure move, no behaviour change) —
expr.rs was already at the src/lower/ line-count ceiling, so even this
guard's few lines needed the headroom.
